final-project/ltt204/data-generation/requirements.txt
Faker==38.2.0
mysql-connector-python==9.5.0
numpy==2.3.5
pandas==2.3.3
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.3
